pyproject.toml
[project]
name = "agent-tools"
version = "0.0.2"
description = "CLI suite for AI-powered development workflows (Jules + GitHub Copilot)"
readme = "README.md"
authors = [
    { name = "copilot-swe-agent[bot]", email = "198982749+Copilot@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.3",
    "httpx>=0.28.1",
    "mcp>=1.27.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
]

[project.scripts]
agent-tools = "agent_tools.cli:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "RET", "SIM", "TID"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

